# Improve HTTP validation for inlining and specialized types