@franz1981 2024-11-17
I've had some fun with assembly and reading the RFC and I'm going to push a major improvement on this, turning this in draft first