@franz1981 2024-11-18
I have pushed https://github.com/eclipse-vertx/vert.x/pull/5392/commits/e4f43420078593445bd8ebe93026b1a1bf82d6f7 and verified it holds from jdk 17 to main.

It solves a problem of "untrusted" types like AsciiString: since the JVM cannot trust offset and end vs arrays boundaries, it would perform additional computations to perform bound checks. 
Giving the JIT an hints that offset is 0 (JIT can profile branches) allow propagating that information, reducing some of these computations in the compiled unrolled version of the loop.
I didn't verify if the same problem happens on end too, since I couldn't spot any weird bookkeeping after this change, although I am sure there is much more to shave.
I would like to produce a code which is less magic 🪄 and still do the right thing.

This explain both why Strings are better performers and why Netty value validation was faster: due to a bug of the Netty version (I will send a pr to fix, since it looks like a potential security issue) the loop variables trigger a better JVM optimization.

I didn't verified