@franz1981 2024-11-22
Any chance you looked into the code @vietj ?