@galderz 2024-12-18
These changes have a positive impact on native.

Before PR:
```
Benchmark                                      (ascii)   Mode  Cnt      Score     Error   Units
HeadersValidationBenchmark.validateNameNetty      true  thrpt    5  13564.945 ± 230.330  ops/ms
HeadersValidationBenchmark.validateNameNetty     false  thrpt    5   4406.927 ±   5.628  ops/ms
HeadersValidationBenchmark.validateNameVertx      true  thrpt    5  17456.963 ±  35.617  ops/ms
HeadersValidationBenchmark.validateNameVertx     false  thrpt    5   3678.795 ±  41.234  ops/ms
HeadersValidationBenchmark.validateValueNetty     true  thrpt    5  16350.225 ± 148.984  ops/ms
HeadersValidationBenchmark.validateValueNetty    false  thrpt    5   3750.169 ±  19.454  ops/ms
HeadersValidationBenchmark.validateValueVertx     true  thrpt    5   1889.591 ±   4.780  ops/ms
HeadersValidationBenchmark.validateValueVertx    false  thrpt    5   2189.690 ± 296.273  ops/ms
```

After PR:
```
Benchmark                                      (ascii)   Mode  Cnt      Score     Error   Units
HeadersValidationBenchmark.validateNameNetty      true  thrpt    5  13696.625 ± 111.232  ops/ms
HeadersValidationBenchmark.validateNameNetty     false  thrpt    5   4234.122 ±  49.438  ops/ms
HeadersValidationBenchmark.validateNameVertx      true  thrpt    5  17824.119 ±  52.503  ops/ms
HeadersValidationBenchmark.validateNameVertx     false  thrpt    5   4028.792 ±  23.738  ops/ms
HeadersValidationBenchmark.validateValueNetty     true  thrpt    5  14507.255 ± 455.145  ops/ms
HeadersValidationBenchmark.validateValueNetty    false  thrpt    5   3515.713 ±  30.857  ops/ms
HeadersValidationBenchmark.validateValueVertx     true  thrpt    5  18613.015 ± 227.081  ops/ms
HeadersValidationBenchmark.validateValueVertx    false  thrpt    5   3505.120 ±  10.858  ops/ms
```

The big winner is `HeadersValidationBenchmark.validateValueVertx` when ascii=true. The reason the performance improves so much is because it produces assembly with no call instructions for this particular benchmark. Some smaller improvements seen in other cases. No regressions.